+++ database/migrations/2020_01_03_170000_update_logins_and_devices_table_user_relation.php
@@ -1,0 +1,30 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class UpdateLoginsAndDevicesTableUserRelation extends Migration
+{
+    public function up()
+    {
+        Schema::table('logins', function (Blueprint $table) {
+            $table->string('user_type')->after('user_id');
+        });
+
+        Schema::table('devices', function (Blueprint $table) {
+            $table->string('user_type')->after('user_id');
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('logins', function (Blueprint $table) {
+            $table->dropColumn('user_type');
+        });
+
+        Schema::table('devices', function (Blueprint $table) {
+            $table->dropColumn('user_type');
+        });
+    }
+}